+++ src/pages/task/finishTask/IndexFinishTask.tsx
@@ -330,7 +330,7 @@ export default function IndexFinishTask() {
             allowClear
           >
             {FINISH_TASK_SEARCH.map((item) => (
-              <option value={item.value}>{item.name}</option>
+              <option key={item.value} value={item.value}>{item.name}</option>
             ))}
           </Select>
         </div>